+++ handy_functions.py
@@ -1,0 +1,33 @@
+import re
+import os
+
+
+def extract_string_between_patterns(
+    input_string, start_pattern, end_pattern
+):  # 提取位于两个字符串中间的特定字符
+    pattern = re.compile(f"{re.escape(start_pattern)}(.*?){re.escape(end_pattern)}")
+    match = pattern.search(input_string)
+    if match:
+        return match.group(1)
+    else:
+        return None
+
+
+def has_subfolder(folder):  # 判断文件夹内是否存在子文件夹
+    names = os.listdir(folder)
+    for name in names:
+        path = os.path.join(folder, name)
+        if os.path.isdir(path):
+            return True
+    return False
+
+
+def arrange_list(strings):
+    # Filter strings that start with "-_-exP_"
+    matching_strings = [s for s in strings if s[:7] == "-_-exP_"]
+
+    # Combine the sorted matching strings with the remaining strings
+    remaining_strings = [s for s in strings if s not in matching_strings]
+    arranged_list = remaining_strings + matching_strings
+
+    return arranged_list
